auth checks and gates

+++ app/Http/Controllers/PostController.php
@@ -8,6 +8,7 @@ use App\Post;
 use Illuminate\Http\Request;
 use App\Http\Requests;
 use Auth;
+use Gate;
 
 
 class PostController extends Controller {
@@ -19,6 +20,9 @@ class PostController extends Controller {
     }
 
     public function getAdminIndex() {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $posts = Post::orderBy('title', 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
@@ -38,18 +42,35 @@ class PostController extends Controller {
     }
 
     public function getAdminCreate() {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+        
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit($id) {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+        $user = Auth::user();
+        if (!$user) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         $post = Post::find($id);
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
     }
 
     public function getAdminDelete($id) {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
+        if (Gate::denies('manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
@@ -57,13 +78,16 @@ class PostController extends Controller {
     }
 
     public function postAdminCreate(Request $request) {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
         $user = Auth::user();
         if(!$user){
-            return redirect()->back();
+            return redirect()->route('admin.index')->with('info', 'Log in for this action to work!');
         }
         $post = new Post([
             'title' => $request->input('title'),
@@ -76,11 +100,17 @@ class PostController extends Controller {
     }
 
     public function postAdminUpdate(Request $request) {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
         $post = POST::find($request->input('id'));
+        if (Gate::denies( 'manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
